refactor : productinformation패키지 명을 information으로 변경
